New files
these contain the notes form, as well as the calculations. I also had to change the login panel to pass in the userid to send to the forms. 
I also added the notesForm.png form to go into the staffviewassets .... I wasn't sure if I had to upload the whole package to change it